// packages/legacy/src/compiler/types.ts
export interface NoteOnEvent {
  kind: 'note_on'
  time: number  // Absolute seconds
  instrumentId: InstrumentId
  note: NoteName
  velocity: number
  duration: number  // Seconds
  articulation?: Articulation
  tie?: 'start' | 'continue' | 'end'
}

export interface NoteOffEvent {
  kind: 'note_off'
  time: number
  instrumentId: InstrumentId
  note: NoteName
  velocity?: number  // Release velocity (0-1)
}

export interface ControlEvent {
  kind: 'control'
  time: number
  instrumentId: InstrumentId
  controller: number
  value: number
}

export interface TempoEvent {
  kind: 'tempo'
  time: number
  bpm: number
  transitionSeconds?: number
  curve?: TempoCurve  // Easing curve for the transition
}